+++ CHANGELOG.md
@@ -12,7 +12,13 @@
 - jslint - Try to improve parser to be able to parse jquery.js without stopping.
 
 # v2026.7.1-beta
-- jslint - bugfix - Fix false-positive unused-variable when variable is used in function argument-default.
+- jslint-bugfix - Fix false-positive-warning 'Unused aa' in code:
+    ```
+    /*jslint white*/
+    let aa = 0;
+    function bb({cc:aa}) {return aa;}
+    bb();
+    ```
 - jslint - Fix jslint unable to continue parsing 'async aa => 0'.
 - jslint-regression - Fix long-running regression where 'let x = x;' doesn't warn about temporal-dead-zone.
 - jslint-warning - Tighten warning of unused variables to be always on, regardless of module / nodejs mode.